+++ crates/algorithms/ecdsa/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#1241](https://github.com/cryspen/libcrux/pull/1241): 
   - Rename `PrivateKey` -> `SigningKey`, `PublicKey` -> `VerificationKey`
-  - Add key-centric and slice-based signature public APIs
+  - Add key-centric signature public APIs
 
 ## [0.0.4] (2025-11-05)